The beginning
Making a simpler and better maze system for an AI to traverse maze and create it. using two block types, path and wall. when fully implemenmted will make maze a bit more refined in terms of traversable space

THIS IS AN EXPERIMENT. Me and bradley are testing to see if this will make the maze look much nicer and still keep all prior functionality