Very rudimentary rooms
makes a simple 3 room by 3 room maze.
Using a dirty main method in Generator for testing, as well as storing walls as "*" and paths as "O"

Needs to be made more flexible, then actual rooms with holes in walls to create paths between rooms need to be made.

Don't know if this is the best way, so if  this is too dumb, please, replace it