Made row/col and y/x notation consistent and make sense
row is y, how far up the player is in maze
col is x, how far across the player is in maze
I know this is simple stuff, but I've been scratching my head over it for a bit. Finally got everything the right way round. For some dumb reason the array is accessed with [y][x]/[row][col], but it works, it's consistent and I'm not touching that specific bit so long it keeps working