Moved some stuff around and made QuestionUI display
now GameScreen holds a QuestionUI. Player still checks if on a trap. When on trap, JumazyController prompts GameScreen to create a new Question. Currently i havent put in checking if player got question correct. but GameScreen should tell JumazyController if player was correct/incorrect, which then sends data back to the model