Found ideal algo for door generation
also updated medieval skin normal wall
Light can now be toggled between two difference stages
default and expanded, use light.increaseLightSize(bool increase) for this. can be used if we decide to add torches as items. very little extra time was spent on this, so dont sweat if we dont do it